can2040: Reschedule tx if previous tx unexpectedly finished transmit
If a transmitted message is fully sent, but the parsing state does not
detect that transmission for some reason (eg, due to a temporary
failure of the local transceiver) then the transmit queue could get
stuck.  Check for a fully transmitted messages when scheduling the
next transmit message to avoid this issue.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>