can2040: Fix PIO "sync" register overflow check
It was possible for data_state_line_passive() to alternate between
MS_DISCARD and MS_START states such that the register overflow check
failed to detect an overflow.  Avoid going back into the MS_DISCARD
state when in the MS_START to prevent the issue.

Also do a full PIO reset on a counter overflow.  There is no harm in
doing a full PIO reset and doing so reduces the number of test
scenarios.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>