can2040: Check for rx message with same id as tx but different payload
The canbus specs state it is not valid for two transmitters to send a
message to the same id at the same time.  Check for this event and
transition to the "discard" state if it is detected.

This additional checking may improve handling of line errors that
occur at the local transceiver.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>