can2040: Do not use canrx line to alter bit timing of cantx line
It is normal for there to be 100+ nanoseconds of delay between a
change to the cantx line and its reception on the canrx line.  At
higher canbus frequencies this delay could result in significant skew
to the tx bit timing.

Change the PIO tx state machine to transmit bits without
resynchronizing bit timing (the bit to bit timing of transmits will be
based only on the rp2040 clock).

The PIO tx state machine bit conflict sampling has been changed from
phase 26 to 20 (of 32 phases).  This should improve the chance that a
simultaneous transmission from another node is properly detected (now
that bit timing is not synchronized).

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>