can2040: Convert report_is_rx_eof_pending() to report_is_not_in_tx()
The report_state may change during a "matched" event on a successfully
received message, which would result in report_is_rx_eof_pending() not
properly reporting that a short EOF is permissible.

Invert the test to avoid this case - thus requiring the full EOF
whenever a self transmit is in progress.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>